+++ pkg/demoinfocs/sendtables2/quantizedfloat.go
@@ -131,7 +131,7 @@ func (qfd *quantizedFloatDecoder) quantize(val float32) float32 {
 	// which must quantize to 1023 so the round-up flag is removed), making the
 	// decoder read a phantom flag bit and desync the entity stream.
 	//nolint:unconvert
-	i := uint32((val-qfd.Low)*qfd.HighLowMul + 0.5)
+	i := uint32(float32((val-qfd.Low)*qfd.HighLowMul) + 0.5)
 	return qfd.Low + float32((qfd.High-qfd.Low)*float32(float32(i)*qfd.DecMul))
 }
 
@@ -149,7 +149,7 @@ func (qfd *quantizedFloatDecoder) decode(r *reader) float32 {
 		return 0.0
 	}
 
-	return qfd.Low + (qfd.High-qfd.Low)*float32(r.readBits(qfd.Bitcount))*qfd.DecMul
+	return qfd.Low + float32((qfd.High-qfd.Low)*float32(r.readBits(qfd.Bitcount))*qfd.DecMul)
 }
 
 // Creates a new quantized float decoder based on given field
